codegen: statically allocate rest param arrays for internal funcs
this 5 line diff makes array push, unshift and more >100x faster!!

internal funcs like array push uses a rest param to allow for variable argument length. that needs allocating which is very slow. we now allocate that statically instead dynamically for every call (only for internal functions as we know it is not reused).